DEBUG: See seismic_data "input" statements
Only difference between fwi.optimizer.param_groups and alan_optimizer.param_groups
is that the tensor being optimized in fwi is "parameter containing tensor==..." and
the tensor being optimized in alan_optimizer is just the tensor itself.

Make sure that this is what is needed or if we need to unwrap down to the tensor
level with the optimizer. I suspect that we are okay; a torch.nn.Parameter is a subclass
of tensor, and its "forward" function MATTERS in the optimization (since it affects the gradient computation).

+++ misfit_toys/misfit_toys_helpers.py
@@ -171,7 +171,7 @@ def towed_src(
     src_depth,
     d_intra_shot
 ):
-    res = torch.zeros(n_shots, src_per_shot, 2)
+    res = torch.zeros(n_shots, src_per_shot, 2, dtype=torch.long)
     res[:, :, 1] = src_depth
     for i in range(n_shots):
         for j in range(src_per_shot):
